test: migrate hidden input python test unique file and move playwright test into jest
