+++ tests/helpers.py
@@ -12,42 +12,59 @@ from webapp.marketo import MarketoAPI
 from webapp.decorators import handle_api_error
 from canonicalwebteam.flask_base.env import get_flask_env
 
+# Fields that are allowed in a Marketo form payload.
+# Add to this set when intentionally introducing a new hidden field.
+ALLOWED_HIDDEN_FIELDS = frozenset(
+    {
+        "firstname",
+        "lastname",
+        "email",
+        "company",
+        "title",
+        "country",
+        "phone",
+        "comments_from_lead__c",
+        "facebook_click_id__c",
+        "gclid__c",
+        "utm_content",
+        "utm_term",
+        "utm_medium",
+        "utm_source",
+        "utm_campaign",
+        "formid",
+        "returnurl",
+        "thankyoumessage",
+        "preferredlanguage",
+        "insightsrobotics",
+        "iot_newsletters__c",
+        "consent_to_processing__c",
+        "canonicalupdatesoptin",
+        # JS-injected fields added at runtime
+        "user_id",
+        "consent_info",
+        "utms",
+    }
+)
+
+
+def get_marketo_template_files():
+    """
+    Returns template files that contain Marketo forms.
+    """
+    result = []
+    for ext in ("*.html", "*.jinja"):
+        for f in Path("templates").rglob(ext):
+            if "templates/tests" in str(f):
+                continue
+            if "/marketo/submit" in f.read_text():
+                result.append(f)
+    return result
+
 
 class MarketoFormTestCase(unittest.TestCase):
     @classmethod
     def setUpClass(cls):
-        # Fields that are allowed in payload
-        cls.SET_FIELDS = set(
-            {
-                "firstname",
-                "lastname",
-                "email",
-                "company",
-                "title",
-                "country",
-                "phone",
-                "comments_from_lead__c",
-                "facebook_click_id__c",
-                "gclid__c",
-                "utm_content",
-                "utm_term",
-                "utm_medium",
-                "utm_source",
-                "utm_campaign",
-                "formid",
-                "returnurl",
-                "thankyoumessage",
-                "preferredlanguage",
-                "insightsrobotics",
-                "iot_newsletters__c",
-                "consent_to_processing__c",
-                "canonicalupdatesoptin",
-                # JS-injected fields added at runtime
-                "user_id",
-                "consent_info",
-                "utms",
-            }
-        )
+        cls.SET_FIELDS = ALLOWED_HIDDEN_FIELDS
 
         marketo_session = Session()
 
@@ -112,19 +129,6 @@ class MarketoFormTestCase(unittest.TestCase):
             for f in Path("templates").rglob("form-data.json")
             if "templates/tests" not in str(f)
         ]
-
-    def _get_marketo_template_files(self):
-        """
-        Helper function to get template files that contain Marketo forms.
-        """
-        result = []
-        for ext in ("*.html", "*.jinja"):
-            for f in Path("templates").rglob(ext):
-                if "templates/tests" in str(f):
-                    continue
-                if "/marketo/submit" in f.read_text():
-                    result.append(f)
-        return result
 
     def _get_contact_us_files(self):
         """
